backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
django-cors-headers==4.6.0
python-dotenv==1.1.0
requests==2.32.3 